(Dataset) From/to rules for ID mapping

This list contains the valid from/to pairings that can be used with uniprot_map(). See Examples below for how to use this object. Also see from_to_dbs for the more information on each database.

Usage

from_to_rules

Format

An object of class list of length 98.

Source

From/to pairs have been downloaded according to the "Valid from and to databases pairs" section on the UniProt ID Mapping page.

See also

Examples

# Show valid `from` values
names(from_to_rules)
#>  [1] "Allergome"                     "ArachnoServer"                
#>  [3] "Araport"                       "BioCyc"                       
#>  [5] "BioGRID"                       "BioMuta"                      
#>  [7] "CCDS"                          "CGD"                          
#>  [9] "ChEMBL"                        "ChiTaRS"                      
#> [11] "CLAE"                          "CollecTF"                     
#> [13] "ComplexPortal"                 "ConoServer"                   
#> [15] "CPTAC"                         "CRC64"                        
#> [17] "dictyBase"                     "DIP"                          
#> [19] "DisProt"                       "DMDM"                         
#> [21] "DNASU"                         "DrugBank"                     
#> [23] "EchoBASE"                      "eggNOG"                       
#> [25] "EMBL-GenBank-DDBJ"             "EMBL-GenBank-DDBJ_CDS"        
#> [27] "Ensembl"                       "Ensembl_Genomes"              
#> [29] "Ensembl_Genomes_Protein"       "Ensembl_Genomes_Transcript"   
#> [31] "Ensembl_Protein"               "Ensembl_Transcript"           
#> [33] "ESTHER"                        "euHCVdb"                      
#> [35] "FlyBase"                       "Gene_Name"                    
#> [37] "GeneCards"                     "GeneID"                       
#> [39] "GeneReviews"                   "GeneTree"                     
#> [41] "GeneWiki"                      "GenomeRNAi"                   
#> [43] "GI_number"                     "GlyConnect"                   
#> [45] "GuidetoPHARMACOLOGY"           "HGNC"                         
#> [47] "HOGENOM"                       "IDEAL"                        
#> [49] "KEGG"                          "LegioList"                    
#> [51] "Leproma"                       "MaizeGDB"                     
#> [53] "MEROPS"                        "MGI"                          
#> [55] "MIM"                           "neXtProt"                     
#> [57] "OMA"                           "OpenTargets"                  
#> [59] "Orphanet"                      "OrthoDB"                      
#> [61] "PATRIC"                        "PDB"                          
#> [63] "PeroxiBase"                    "PharmGKB"                     
#> [65] "PHI-base"                      "PIR"                          
#> [67] "PlantReactome"                 "PomBase"                      
#> [69] "ProteomicsDB"                  "PseudoCAP"                    
#> [71] "Reactome"                      "REBASE"                       
#> [73] "RefSeq_Nucleotide"             "RefSeq_Protein"               
#> [75] "RGD"                           "SGD"                          
#> [77] "STRING"                        "SwissLipids"                  
#> [79] "TCDB"                          "TreeFam"                      
#> [81] "TubercuList"                   "UCSC"                         
#> [83] "UniParc"                       "UniPathway"                   
#> [85] "UniProtKB_AC-ID"               "UniRef100"                    
#> [87] "UniRef50"                      "UniRef90"                     
#> [89] "VEuPathDB"                     "VGNC"                         
#> [91] "WBParaSite"                    "WBParaSite_Transcript-Protein"
#> [93] "World-2DPAGE"                  "WormBase"                     
#> [95] "WormBase_Protein"              "WormBase_Transcript"          
#> [97] "Xenbase"                       "ZFIN"                         

# Show valid `to` values for a given `from` e.g. SGD
from_to_rules[["SGD"]]
#> [1] "UniProtKB"            "UniProtKB-Swiss-Prot"