(Dataset) UniProt return fields

This dataframe contains the valid fields (i.e. columns) of data you can request from UniProt. The strings in the field column is what you'll use with the functions in this package. See Examples below for how to use this object.

Usage

return_fields

Format

An object of class data.frame with 389 rows and 4 columns.

Source

UniProtKB return fields have been scraped from the UniProtKB return fields page. The return fields from other Uniprot databases have been determined by hand using Web Developer Tools (F12) to inspect the GET request made when searching the different database on the UniProt website.

Details

Columns:

database	character, UniProt database to be queried.

section	character, similar return fields are grouped together in sections.

field	character, the return field i.e. the string used to request the desired column of information.

label	character, human readable column name that will be returned by the API.

See also

Examples

# What UniProt databases are available to query?
levels(return_fields$database)
#> [1] "uniprotkb" "uniref"    "uniparc"   "proteomes" "taxonomy"  "keywords" 
#> [7] "citations" "diseases"  "locations"

# What fields are available for the `proteomes` database?
return_fields[return_fields$database == "proteomes", "field"]
#>  [1] "upid"                  "organism"              "organism_id"          
#>  [4] "components"            "mnemonic"              "lineage"              
#>  [7] "busco"                 "cpd"                   "genome_assembly"      
#> [10] "genome_representation" "protein_count"